server/tennis_app/requirements.txt
asgiref==3.7.2
Django==3.2.13
django-admin-list-filters==1.1
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
django-cors-headers1
